Bug fix on MaxVMBuildOP and Mod 7 scripts

+++ DIY/Maximus/Scripts/7Remove.ps1
@@ -1,27 +1,30 @@
-$KillGateway = $false
-$KillOPVM    = $true
-$KillOPNVA   = $false
-$KillCSVM    = $false
-$KillAll     = $false
+$KillGateway  = $false  # Remove things related to the gateway
+$KillOPVM     = $true   # Remove things related to the OnPrem VM
+$KillOPNVA    = $false  # Remove things related to the OnPrem Router
+$KillCSVM     = $false  # Remove things related to the Coffee Shop VM
+$KillAll      = $false  # Remove all of the above, plus secrets and files
+$KillComplete = $false  # Removes everything created in Mod 7, even VNets and Bastion servers
 
-if ($KillAll) {
+if ($KillAll -or $KillComplete) {
     $KillGateway = $true
-    $KillOPVM = $true
-    $KillOPNVA = $true
-    $KillCSVM = $true
+    $KillOPVM    = $true
+    $KillOPNVA   = $true
+    $KillCSVM    = $true
+    $KillAll     = $true
 }
 
 $RGName   = "MaxLab"
-$kvName   = $kvName = (Get-AzKeyVault -ResourceGroupName $RGName | Select-Object -First 1).VaultName
+$kvName   = (Get-AzKeyVault -ResourceGroupName $RGName | Select-Object -First 1).VaultName
+$SAName   = (Get-AzStorageAccount -ResourceGroupName $RGName | Select-Object -First 1).StorageAccountName
 $HubName  = "Hub-VNet"
 $OPName   = "OnPrem-VNet"
 $OPVMName = "OnPrem-VM01"
 $CSVMName = "CoffeeShop-PC"
-$S1Name = "Spoke01-VNet"
-$S2Name = "Spoke02-Vnet"
+$S1Name   = "Spoke01-VNet"
+$S2Name   = "Spoke02-Vnet"
 
-if ($KillGW) {
-    Write-Host "Killing GW Connections"
+if ($KillGateway) {
+    Write-Host "Killing GW Connection"
     try {Get-AzVirtualNetworkGatewayConnection -ResourceGroupName $RGName -Name $HubName-gw-op-conn -ErrorAction Stop | Out-Null
          Remove-AzVirtualNetworkGatewayConnection -ResourceGroupName $RGName -Name $HubName-gw-op-conn -Force}
     catch {Write-Host "  It's not there"}
@@ -56,7 +59,18 @@ foreach ($VMName in $VMNames) {
     catch {Write-Host "  It's not there"}    
 }
 
-if ($KillAll) {
+if ($KillComplete) {
+    $Bastions = @()
+    $Bastions += $OPName + '-bas'
+    $Bastions += $CSName + '-bas'
+    foreach ($Bastion in $Bastions) {
+        Write-Host "Killing $Bastion"
+        try {Get-AzBastion -ResourceGroupName $RGName -Name $Bastion -ErrorAction Stop | Remove-AzBastion -Force -AsJob}
+        catch {Write-Host "  It's not there"}
+    }
+}
+
+If ($KillOPVM){
     $SecretNames = @()
     $SecretNames += "OnPremNVArsa"
     $SecretNames += "S2SPSK"
@@ -76,9 +90,24 @@ if ($KillAll) {
         Else {Remove-AzKeyVaultSecret -VaultName $kvName -Name $SecretName -InRemovedState -Force}
     }
 
+    Write-Host "Killing Client.pfx from Storage"
+    $sa = Get-AzStorageAccount -ResourceGroupName $RGName -Name $SAName -ErrorAction Stop
+    try {Get-AzStorageBlob -Container $web -Blob "Client.pfx" -Context $sa.Context -ErrorAction Stop| Remove-AzStorageBlob -Force}
+    catch {Write-Host "  It's not there"}
+
+    Write-Host "Killing config container from Storage"
+    try {Get-AzStorageContainer -Container "config" -Context $sa.Context -ErrorAction Stop| Remove-AzStorageContainer -Force}
+    catch {Write-Host "  It's not there"}
+}
+
+if ($KillAll) {
     Write-Host "Killing Local S2S GW"
     try {Get-AzLocalNetworkGateway -Name $OPName'-lgw' -ResourceGroupName $RGName -ErrorAction Stop | Out-Null
         Remove-AzLocalNetworkGateway -Name $OPName'-lgw' -ResourceGroupName $RGName -Force -AsJob}
+    catch {Write-Host "  It's not there"}
+
+    Write-Host "Killing OnPrem Route Table"
+    try {Get-AzRouteTable -ResourceGroupName $RGName -Name $OPName'-rt' -ErrorAction Stop | Remove-AzRouteTable -Force}
     catch {Write-Host "  It's not there"}
 
     Write-Host "Resetting Spoke VNets peering"
@@ -116,10 +145,20 @@ foreach ($VMName in $VMNames) {
 Write-Host "Waiting for NICs to delete"
 Get-Job -Command "Remove-AzNetworkInterface" | Wait-Job -Timeout 600 | Out-Null
 
+if ($KillComplete) {
+    Write-Host "Waiting for the Bastions to delete"
+    Get-Job -Command "Remove-AzBastion" | Wait-Job -Timeout 600 | Out-Null
+    foreach ($Bastion in $Bastions) {
+        Write-Host "Killing $Bastion PIP"
+        try {Get-AzPublicIPAddress -ResourceGroupName $RGName -Name $Bastion"-pip" -ErrorAction Stop | Remove-AzPublicIPAddress -Force -AsJob}
+        catch {Write-Host "  It's not there"}
+    }    
+}
+
 if ($KillOPNVA) {
     Write-Host "Killing $VMName PIP"
     try {Get-AzPublicIPAddress -ResourceGroupName $RGName -Name $VMName"-pip" -ErrorAction Stop | Out-Null
-        Remove-AzPublicIPAddress -ResourceGroupName $RGName -Name $VMName"-pip" -Force -AsJob}
+         Remove-AzPublicIPAddress -ResourceGroupName $RGName -Name $VMName"-pip" -Force -AsJob}
     catch {Write-Host "  It's not there"}    
 }
